Improved trainer
The trainer now has a separate step() function which is called inside
the while loop of run(), allowing steps to be made separately if need
be.

The “dynamic” selection method now does a bit more to try and overcome
local maxima, and also can determine if it’s close to the absolute max
and continue normally.